@@ -1,172 +1,117 @@
-import 'dart:async';
-import 'dart:math';
-
-import 'package:dart_emoji/dart_emoji.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_data/flutter_data.dart';
-import 'package:intl/intl.dart';
-import 'package:toastification/toastification.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:models/models.dart';
+import 'package:zapstore/constants/app_constants.dart';
+import 'package:zapstore/services/package_manager/package_manager.dart';
+import 'package:zapstore/utils/version_utils.dart';
+import 'package:collection/collection.dart';
 
-extension ContextX on BuildContext {
-  ThemeData get theme => Theme.of(this);
-  void showInfo(String message, {String? description, Icon? icon}) {
-    toastification.show(
-      context: this,
-      type: ToastificationType.info,
-      icon: icon ?? Icon(Icons.info),
-      style: ToastificationStyle.fillColored,
-      alignment: Alignment.topCenter,
-      title: _ToastTitleWidget(message),
-      description: description != null
-          ? _ToastDescriptionWidget(description: description)
-          : null,
-      autoCloseDuration: const Duration(seconds: 4),
-      showProgressBar: false,
-      closeOnClick: true,
-    );
-  }
+// Re-export constants for backwards compatibility
+export 'package:zapstore/constants/app_constants.dart';
 
-  void showError(String title,
-      {String? description,
-      Icon? icon,
-      List<(String, Future<void> Function())> actions = const []}) {
-    toastification.show(
-      context: this,
-      type: ToastificationType.error,
-      style: ToastificationStyle.fillColored,
-      alignment: Alignment.topCenter,
-      icon: icon ?? Icon(Icons.error),
-      title: _ToastTitleWidget(title),
-      showProgressBar: false,
-      closeOnClick: true,
-      description: description != null
-          ? _ToastDescriptionWidget(
-              description: description,
-              actions: actions,
-            )
-          : null,
-    );
-  }
+extension WidgetExt on WidgetRef {
+  StorageNotifier get storage => read(storageNotifierProvider.notifier);
+  PackageManager get packageManager => read(packageManagerProvider.notifier);
 }
 
-class _ToastTitleWidget extends StatelessWidget {
-  final String text;
-  const _ToastTitleWidget(this.text);
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      text,
-      style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-      overflow: TextOverflow.ellipsis,
-      maxLines: 4,
-    );
-  }
+extension ContextExt on BuildContext {
+  TextTheme get textTheme => Theme.of(this).textTheme;
 }
 
-class _ToastDescriptionWidget extends StatelessWidget {
-  final String? description;
-  final List<(String, Future<void> Function())> actions;
+extension AppExt on App {
+  bool get isRelaySigned =>
+      author.value?.pubkey == kZapstorePubkey &&
+      identifier != kZapstoreAppIdentifier;
 
-  _ToastDescriptionWidget({this.description, this.actions = const []});
+  /// Returns PackageInfo if installed, otherwise null
+  PackageInfo? get installedPackage =>
+      ref.read(packageManagerProvider.notifier).getInfo(identifier);
 
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        RichText(
-          text: TextSpan(
-            text: description,
-            style: TextStyle(fontSize: 16),
-          ),
-        ),
-        for (final (text, fn) in actions)
-          ElevatedButton(
-            style: ElevatedButton.styleFrom(
-                foregroundColor: Colors.white,
-                backgroundColor: Colors.transparent),
-            onPressed: () async {
-              await fn.call();
-              toastification.dismissAll(delayForAnimation: false);
-            },
-            child: Text(text),
-          ),
-      ],
-    );
-  }
-}
+  /// Whether the app is installed on the device
+  bool get isInstalled =>
+      ref.read(packageManagerProvider.notifier).isInstalled(identifier);
 
-BelongsTo<T> belongsTo<T extends DataModelMixin<T>>(Map<String, dynamic>? map) {
-  return map != null ? BelongsTo<T>.fromJson(map) : BelongsTo<T>();
-}
+  /// Latest file metadata associated to the latest release
+  /// Note: assumes latest metadata has been loaded for the current platform
+  FileMetadata? get latestFileMetadata =>
+      latestRelease.value?.latestMetadata.value;
 
-HasMany<T> hasMany<T extends DataModelMixin<T>>(Map<String, dynamic>? map) {
-  return map != null ? HasMany<T>.fromJson(map) : HasMany<T>();
-}
+  /// Whether there is an update available for the installed app
+  /// Compares versionCode first when available, otherwise falls back to
+  /// semantic version comparison.
+  bool get hasUpdate {
+    final installed = installedPackage;
+    final latest = latestFileMetadata;
+    if (installed == null || latest == null) return false;
 
-final emojiParser = EmojiParser();
-
-extension StringWidget on String {
-  String parseEmojis() {
-    return replaceAllMapped(RegExp(':([a-z]*):'), (m) {
-      return emojiParser.hasName(m[1]!) ? emojiParser.get(m[1]!).code : m[0]!;
-    });
-  }
-
-  String safeSubstring(int size) {
-    return substring(0, min(length, size)) + (size < length ? '...' : '');
-  }
-
-  String get shorten {
-    // npub1wf4puf...43dgh9
-    if (length < 18) return this;
-    final leading = substring(0, 9);
-    final trailing = substring(length - 6, length - 1);
-    return '$leading...$trailing';
-  }
-
-  String removeParenthesis() {
-    return replaceAll(RegExp(r'\([^()]*\)'), '');
-  }
-}
-
-extension TextExt on Text {
-  Widget get bold {
-    return Text(data!, style: TextStyle(fontWeight: FontWeight.bold));
-  }
-}
-
-const kZapstoreAppIdentifier = 'dev.zapstore.app';
-
-final kNumberFormatter = NumberFormat('#,###');
-
-// stream utils
-
-Stream<List<T>> bufferByTime<T>(Stream<T> source, Duration duration) {
-  final controller = StreamController<List<T>>();
-
-  final buffer = [];
-
-  Timer? timer;
-
-  // Listen to the source stream
-  source.listen((data) {
-    buffer.add(data);
-
-    // If there's no active timer, start one
-    timer ??= Timer(duration, () {
-      controller.add(List.from(buffer)); // Emit buffered data
-      buffer.clear(); // Clear the buffer
-      timer = null; // Reset the timer
-    });
-  }, onDone: () {
-    // Emit remaining items in the buffer when the source is done
-    if (buffer.isNotEmpty) {
-      controller.add(List.from(buffer));
+    if (latest.versionCode != null && installed.versionCode != null) {
+      return latest.versionCode! > installed.versionCode!;
     }
-    controller.close(); // Close the controller
-  });
 
-  return controller.stream; // Return the buffered stream
+    return canUpgrade(installed.version, latest.version);
+  }
+
+  /// Whether the relay version would be a downgrade from the installed version
+  /// Compares versionCode first when available, otherwise falls back to
+  /// semantic version comparison.
+  bool get hasDowngrade {
+    final installed = installedPackage;
+    final latest = latestFileMetadata;
+    if (installed == null || latest == null) return false;
+
+    if (latest.versionCode != null && installed.versionCode != null) {
+      return latest.versionCode! < installed.versionCode!;
+    }
+
+    return canUpgrade(latest.version, installed.version);
+  }
+
+  /// Whether the installed app is up to date (installed and no update)
+  bool get isUpdated => isInstalled && !hasUpdate;
+}
+
+extension AppsExt on Iterable<App> {
+  Future<void> loadMetadata({bool withAuthors = true}) async {
+    if (isEmpty) return;
+    final ref = first.ref;
+
+    final releases = await ref.storage.query(
+      Request(
+        map(
+          (app) => app.latestRelease.req?.filters.firstOrNull,
+        ).nonNulls.toList(),
+      ),
+    );
+
+    if (releases.isEmpty) return;
+
+    await ref.storage.query(
+      Request<FileMetadata>(
+        releases
+            .map((r) => r.latestMetadata.req?.filters.firstOrNull)
+            .nonNulls
+            .toList(),
+      ),
+      source: const LocalAndRemoteSource(stream: false, background: true),
+    );
+
+    if (withAuthors) {
+      await ref.storage.query(
+        Request<Profile>(
+          map(
+            (a) => a.author.req?.filters.firstOrNull,
+          ).nonNulls.toSet().toList(),
+        ),
+        source: const LocalAndRemoteSource(
+          relays: 'vertex',
+          stream: false,
+          background: true,
+        ),
+      );
+    }
+  }
+}
+
+extension WidgetRefExt on WidgetRef {
+  Ref get ref => read(Provider((ref) => ref));
 }
