@@ -1,82 +1,196 @@
 import 'package:flutter/material.dart';
-import 'package:gap/gap.dart';
-import 'package:zapstore/models/app.dart';
-import 'package:zapstore/models/local_app.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:models/models.dart';
 import 'package:zapstore/utils/extensions.dart';
-import 'package:zapstore/widgets/install_button.dart';
-import 'package:zapstore/widgets/pill_widget.dart';
+import '../services/package_manager/package_manager.dart';
+import '../theme.dart';
 
-class VersionPillWidget extends StatelessWidget {
+/// Version pill widget showing the current release version
+/// Displays version in a colored pill format similar to the old design
+/// Can also show dual versions (installed vs available) for updates
+class VersionPillWidget extends HookConsumerWidget {
+  final App app;
+  final bool showUpdateArrow;
+  final String? forceVersion;
+
   const VersionPillWidget({
     super.key,
     required this.app,
+    this.showUpdateArrow = false,
+    this.forceVersion,
   });
 
-  final App app;
-
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
+    // Watch package manager state for reactivity
+    final installedPackages = ref.watch(packageManagerProvider);
+
+    // Resolve installed and available versions using AppExt
+    final installedVersion = app.installedPackage?.version;
+    final availableVersion = app.latestFileMetadata?.version;
+    final updateAvailable = app.hasUpdate;
+    final downgradeAvailable = app.hasDowngrade;
+
+    // Dual version mode (when an update OR downgrade is available)
+    if (showUpdateArrow &&
+        (updateAvailable || downgradeAvailable) &&
+        installedVersion != null &&
+        availableVersion != null) {
+      return _buildDualVersionPills(
+        context,
+        ref,
+        installedPackages,
+        installedVersion,
+        availableVersion,
+        isDowngrade: downgradeAvailable,
+      );
+    }
+
+    // Single version mode: use forced version if provided, otherwise show installed version if installed, otherwise available version
+    final version = forceVersion ?? installedVersion ?? availableVersion;
+
+    if (version == null || version.isEmpty) {
+      return const SizedBox.shrink();
+    }
+
+    return _buildVersionPill(
+      context,
+      ref,
+      installedPackages,
+      version,
+      AppColors.darkPillBackground,
+      Colors.white,
+    );
+  }
+
+  Widget _buildDualVersionPills(
+    BuildContext context,
+    WidgetRef ref,
+    List<PackageInfo> installedPackages,
+    String installedVersion,
+    String availableVersion, {
+    bool isDowngrade = false,
+  }) {
     return Row(
+      mainAxisSize: MainAxisSize.min,
       children: [
-        if (app.canUpdate)
-          PillWidget(
-            WidgetSpan(
-              alignment: PlaceholderAlignment.middle,
-              child: Wrap(
-                crossAxisAlignment: WrapCrossAlignment.center,
-                children: [
-                  Text(
-                    app.localApp.value!.installedVersion!.removeParenthesis(),
-                    style: TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 11,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            size: 10,
-            color: Colors.grey[800]!,
-          ),
-        if (app.canUpdate) Icon(Icons.arrow_right),
-        if (app.latestMetadata != null)
-          PillWidget(
-            WidgetSpan(
-              alignment: PlaceholderAlignment.middle,
-              child: Wrap(
-                crossAxisAlignment: WrapCrossAlignment.center,
-                children: [
-                  Text(
-                    app.latestMetadata!.version!.removeParenthesis(),
-                    style: TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 11,
-                    ),
-                  ),
-                  Row(
-                    children: [
-                      Gap(5),
-                      if (app.canUpdate) Icon(Icons.update_outlined, size: 15),
-                      if (app.canInstall)
-                        Icon(Icons.download_rounded, size: 15),
-                      if (app.isUpdated) Icon(Icons.check, size: 15),
-                      if (app.isDowngrade) Icon(Icons.do_not_disturb, size: 15),
-                      if (app.hasCertificateMismatch)
-                        Icon(Icons.do_not_disturb, size: 15),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-            size: 10,
-            color: switch (app.localApp.value?.status) {
-              AppInstallStatus.downgrade ||
-              AppInstallStatus.certificateMismatch =>
-                Colors.grey[800]!,
-              _ => kUpdateColor,
-            },
-          ),
+        // Current version pill (muted colors for installed version)
+        _buildVersionPill(
+          context,
+          ref,
+          installedPackages,
+          installedVersion,
+          Theme.of(context).colorScheme.outline.withValues(alpha: 0.3),
+          Theme.of(context).colorScheme.onSurface,
+          isInstalledVersion: true,
+        ),
+
+        // Arrow icon (always arrow, forbidden icon is in the pill itself)
+        Icon(
+          Icons.arrow_right,
+          size: 16,
+          color: Theme.of(context).colorScheme.onSurfaceVariant,
+        ),
+
+        // Available version pill (highlighted for update, greyed for downgrade)
+        _buildVersionPill(
+          context,
+          ref,
+          installedPackages,
+          availableVersion,
+          isDowngrade
+              ? Theme.of(context).colorScheme.outline.withValues(alpha: 0.3)
+              : AppColors.darkPillBackground,
+          isDowngrade
+              ? Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.5)
+              : Colors.white,
+          isDowngrade: isDowngrade,
+        ),
       ],
     );
   }
+
+  Widget _buildVersionPill(
+    BuildContext context,
+    WidgetRef ref,
+    List<PackageInfo> installedPackages,
+    String version,
+    Color backgroundColor,
+    Color textColor, {
+    bool isInstalledVersion = false,
+    bool isDowngrade = false,
+  }) {
+    // Trim version text to max 8 characters to make room for icon
+    String displayVersion = _displayVersion(version);
+
+    // Determine app status for icon
+    Widget? statusIcon;
+    if (!isInstalledVersion) {
+      if (isDowngrade) {
+        // Downgrade forbidden
+        statusIcon = Icon(Icons.block, size: 12, color: textColor);
+      } else if (!app.isInstalled) {
+        // Can install
+        statusIcon = const Icon(
+          Icons.download_rounded,
+          size: 12,
+          color: Colors.white,
+        );
+      } else {
+        if (app.hasUpdate) {
+          // Can update
+          statusIcon = const Icon(
+            Icons.update_outlined,
+            size: 12,
+            color: Colors.white,
+          );
+        } else {
+          // Is updated
+          statusIcon = const Icon(Icons.check, size: 12, color: Colors.white);
+        }
+      }
+    }
+
+    // Use grey colors for "do not disturb" states (downgrade, certificate mismatch)
+    final finalBackgroundColor = backgroundColor;
+    final finalTextColor = textColor;
+
+    final double verticalPadding = isInstalledVersion ? 5.0 * 1.05 : 5.0;
+
+    return Container(
+      padding: EdgeInsets.symmetric(horizontal: 9, vertical: verticalPadding),
+      decoration: BoxDecoration(
+        color: finalBackgroundColor,
+        borderRadius: BorderRadius.circular(8),
+      ),
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text(
+            displayVersion,
+            maxLines: 1,
+            overflow: TextOverflow.ellipsis,
+            style: context.textTheme.labelMedium?.copyWith(
+              fontWeight: FontWeight.bold,
+              color: finalTextColor,
+              fontSize: 11.5,
+            ),
+          ),
+          if (statusIcon != null) ...[
+            const SizedBox(width: 4),
+            IconTheme.merge(
+              data: const IconThemeData(size: 12),
+              child: statusIcon,
+            ),
+          ],
+        ],
+      ),
+    );
+  }
+
+  String _displayVersion(String version) {
+    return version.length > 10 ? '${version.substring(0, 9)}...' : version;
+  }
+
+  // No width enforcement; both pills share the same text style
 }
